Fix build issue
