+++ package.json
@@ -10,6 +10,7 @@
 	},
 	"dependencies": {
 		"next": "^13.4.3",
+		"nextra": "^2.10.0",
 		"nextra-theme-docs": "^2.6.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0"
