+++ pages/blog.mdx
@@ -3,8 +3,8 @@ description: The latest updates and releases from the team at Allo.
 type: posts
 ---
 
-import { BlogHeader } from "../components/blog/BlogHeader";
-import { BlogIndex } from "../components/blog/BlogIndex";
+import { BlogHeader } from "../components/BlogHeader";
+import { BlogIndex } from "../components/BlogIndex";
 
 <BlogHeader />
 <BlogIndex />
